fixed entrypoint.sh to handle non-interactive shell

+++ build/docker/entrypoint.sh
@@ -2,7 +2,8 @@
 
 graceful_shutdown() {
   PID="$(ps -e -o pid,cmd | awk '/serverjockey\.pyz/{print $1}')"
-  echo "Trapped shutdown signal, terminating $PID"
+  [ -z "$PID" ] && exit 0
+  echo "Graceful shutdown, terminating $PID"
   kill $PID
   while ps -p $PID > /dev/null; do sleep 1; done
   exit 0
@@ -12,19 +13,16 @@ sleep 1
 TZ=${TZ:-UTC}
 export TZ
 cd /home/container || exit 1
-if [ ! -d "serverlink" ]; then
-  mkdir serverlink || exit 1
-  echo '{ "module": "serverlink", "hidden": true }' > serverlink/instance.json
-  /usr/games/steamcmd +quit
-fi
-
+[ -d ".local/share/Steam" ] || /usr/games/steamcmd +quit
 [ -z "$STARTUP" ] && STARTUP="/usr/local/bin/serverjockey.pyz --noupnp --showtoken"
 MODIFIED_STARTUP=$(echo ${STARTUP} | sed -e 's/{{/${/g' -e 's/}}/}/g')
 echo -e ":/home/container$ ${MODIFIED_STARTUP}"
 trap "graceful_shutdown" TERM INT
 eval ${MODIFIED_STARTUP} &
+WAIT_PID=$!
 while read -r line; do
-  [ "$line" = "shutdown" ] && break
-  eval "/usr/local/bin/serverjockey_cmd.pyz -u container -c $line"
+  [ "$line" = "shutdown" ] && graceful_shutdown
+  eval /usr/local/bin/serverjockey_cmd.pyz -u container -c $line
 done
-graceful_shutdown
+wait $WAIT_PID
+exit 0
